Split out formatPrice
