+++ client/lib/gsuite/index.js
@@ -1,7 +1,6 @@
 /**
  * External dependencies
  */
-import formatCurrency from '@automattic/format-currency';
 import { endsWith, get, isNumber, isString, sortBy, includes } from 'lodash';
 
 /**
@@ -12,20 +11,9 @@ import {
 	GSUITE_BUSINESS_SLUG,
 	GSUITE_EXTRA_LICENSE_SLUG,
 } from 'lib/gsuite/constants';
+import { formatPrice } from 'lib/gsuite/utils/format-price';
 import { isMappedDomainWithWpcomNameservers, isRegisteredDomain } from 'lib/domains';
 import userFactory from 'lib/user';
-
-/**
- * Applies a precision to the cost
- *
- * @param {number} cost - cost
- * @param {number} precision - precision to apply to cost
- * @returns {string} - Returns price with applied precision
- */
-function applyPrecision( cost, precision ) {
-	const exponent = Math.pow( 10, precision );
-	return Math.ceil( cost * exponent ) / exponent;
-}
 
 /**
  * Determines whether G Suite is allowed for the specified domain.
@@ -39,22 +27,6 @@ export function canDomainAddGSuite( domainName ) {
 	}
 
 	return canUserPurchaseGSuite();
-}
-
-/**
- * Formats price given cost and currency
- *
- * @param {number} cost - cost
- * @param {string} currencyCode - currency code to format with
- * @param {object} options - options containing precision
- * @returns {string} - Returns a formatted price
- */
-function formatPrice( cost, currencyCode, options = {} ) {
-	if ( undefined !== options.precision ) {
-		cost = applyPrecision( cost, options.precision );
-	}
-
-	return formatCurrency( cost, currencyCode, cost % 1 > 0 ? {} : { precision: 0 } );
 }
 
 /**
